add solution for MJT hw1

+++ MJT/hw1/src/bg/sofia/uni/fmi/mjt/glovo/delivery/DeliveryType.java
@@ -1,18 +1,18 @@
 package bg.sofia.uni.fmi.mjt.glovo.delivery;
 
 public enum DeliveryType {
-    CAR(5, 3),
-    BIKE(3, 5);
+    CAR(5.0, 3),
+    BIKE(3.0, 5);
 
-    private final int pricePerKilometer;
+    private final double pricePerKilometer;
     private final int timePerKilometer;
 
-    DeliveryType(int pricePerKilometer, int timePerKilometer) {
+    DeliveryType(double pricePerKilometer, int timePerKilometer) {
         this.pricePerKilometer = pricePerKilometer;
         this.timePerKilometer = timePerKilometer;
     }
 
-    public int getPricePerKilometer() {
+    public double getPricePerKilometer() {
         return pricePerKilometer;
     }
